Report also on number of unacknowleged messages
The unacknowledged messages gives an accurate image of what the actual status is of the number tasks currently running on the queue.
This information can used by the chunker worker to see if can be scaled up.
For example when a new task is added to the queue it calculates how many workers could be used based on the amount of workers available.
But if there are already tasks on the queue it could result in overshooting.
i.e. too many messages on the queue because the new work will take what it thinks is can use, but that is not yet available.
As a results to many tasks will be on the queue and the queue will need to time to get ride of that number of messages.
Because the system still uses the task load and thus the overshoot keeps in place until tasks are finished.
It's not a problem for the operations itself, it's just the load on the queue becomes higher than actually needed.